pyproject.toml
[project]
name = "pft"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.12"
dependencies = [
    "google-adk>=1.26.0",
    "python-dotenv>=1.2.2",
]
